requirements.txt
torch==2.0.1
transformers==4.30.2
datasets==2.14.4
networkx==3.1
scikit-learn==1.3.0
gradio
faiss-gpu